+++ android/src/main/kotlin/m3u8_player/M3u8Player.kt
@@ -1,0 +1,26 @@
+package m3u8_player
+
+import androidx.annotation.NonNull
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.plugin.common.MethodCall
+import io.flutter.plugin.common.MethodChannel
+import io.flutter.plugin.common.MethodChannel.MethodCallHandler
+
+/** Minimal implementation for the M3u8Player plugin on Android */
+class M3u8Player : FlutterPlugin, MethodCallHandler {
+  private lateinit var channel: MethodChannel
+
+  override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+    channel = MethodChannel(flutterPluginBinding.binaryMessenger, "m3u8_player")
+    channel.setMethodCallHandler(this)
+  }
+
+  override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
+    // This plugin does not implement any native functionality.
+    result.notImplemented()
+  }
+
+  override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+    channel.setMethodCallHandler(null)
+  }
+}