+++ CHANGELOG.md
@@ -1,3 +1,16 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## [Unreleased] - 2025-03-21
+
+### Git Changelog
+- Implemented core functionalities for m3u8 player.
+- Refactored player control components.
+- Improved quality selection services.
+- Minor bug fixes and performance improvements.
+
+### pub.dev Changelog
+- v0.0.3 Release:
+    - Implemented core functionalities for m3u8 player.
+    - Refactored player control components.
+    - Improved quality selection services.
+    - Minor bug fixes and performance improvements.